requirements.txt
PyYAML==6.0.1
requests==2.28.2
pandas==2.2.0
numpy==1.23.5
xlrd==1.2.0
TA-Lib==0.4.32
tables==3.9.1
schedule==0.6.0
wxpusher==2.2.0
pytest==7.2.0
akshare==1.14.60
urllib3>=1.26.0,<2.0
openpyxl